requirements.txt
streamlit
numpy==1.22.3
pandas==1.4.2
pillow==9.0.1
